WIP: Initial support for charset plugins
